+++ definition/user-api-tokens/openapi/components/parameters/apiTokenId.@.js
@@ -1,0 +1,9 @@
+export default {
+	description: 'The identifier of the specific API token.',
+	name: 'apiTokenId',
+	in: 'path',
+	required: true,
+	schema: {
+		type: 'string',
+	},
+}