+++ definition/user-api-tokens/openapi/components/parameters/userApiTokenId.@.js
@@ -1,9 +1,0 @@
-export default {
-	description: 'The identifier of the specific API token owned by a single user.',
-	name: 'userApiTokenId',
-	in: 'path',
-	required: true,
-	schema: {
-		type: 'string',
-	},
-}