+++ definition/user-api-tokens/openapi/tags.@.js
@@ -1,4 +1,4 @@
 export const userApiTokens = {
-	description: 'Routes that handle managing API tokens that are used by individual users.',
+	description: 'Routes that handle managing API tokens that are owned by individual users.',
 	name: 'userApiTokens',
 }