+++ definition/user-api-tokens/routes/paths/self/apiTokens/get.@.js
@@ -18,7 +18,7 @@ export const parameters = [
 
 export const responses = {
 	200: {
-		description: 'The filtered list of the users API tokens.',
+		description: 'The filtered list of the API tokens.',
 		content: {
 			'application/json': {
 				schema: {
@@ -27,7 +27,7 @@ export const responses = {
 						data: {
 							type: 'array',
 							items: {
-								$ref: '#/components/schemas/userApiToken',
+								$ref: '#/components/schemas/apiToken',
 							},
 						},
 					},
@@ -41,9 +41,9 @@ export const responses = {
 }
 
 export default async request => {
-	const { userApiTokens } = await request.controller.userApiToken.list(request)
+	const { apiTokens } = await request.controller.apiToken.list(request)
 	return {
 		status: 200,
-		body: { data: userApiTokens },
+		body: { data: apiTokens },
 	}
 }