+++ definition/user-tasks/openapi/components/parameters/taskId.@.js
@@ -1,0 +1,9 @@
+export default {
+	description: 'The identifier of the specific task.',
+	name: 'taskId',
+	in: 'path',
+	required: true,
+	schema: {
+		type: 'string',
+	},
+}