+++ definition/user-tasks/routes/paths/self/taskGroups/post.@.js
@@ -1,0 +1,50 @@
+export const summary = 'Create User Task Group'
+
+export const description = 'Create a task grouping that is owned by a single user.'
+
+export const tags = [ 'userTasks' ]
+
+export const requestBody = {
+	description: 'Create a task grouping.',
+	content: {
+		'application/json': {
+			schema: {
+				type: 'object',
+				properties: {
+					data: {
+						$ref: '#/components/schemas/taskGroup',
+					},
+				},
+			},
+		},
+	},
+}
+
+export const responses = {
+	201: {
+		description: 'The task grouping was created successfully',
+		content: {
+			'application/json': {
+				schema: {
+					type: 'object',
+					properties: {
+						data: {
+							$ref: '#/components/schemas/taskGroup',
+						},
+					},
+				},
+			},
+		},
+	},
+	default: {
+		$ref: '#/components/responses/error',
+	},
+}
+
+export default async request => {
+	const { taskGroup } = await request.controller.taskGroup.create(request)
+	return {
+		status: 201,
+		body: { data: taskGroup },
+	}
+}